+++ backend/apid/middlewares/authorization.go
@@ -1,10 +1,15 @@
 package middlewares
 
 import (
+	"context"
 	"net/http"
 
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/apid/actions"
 	"github.com/sensu/sensu-go/backend/authorization"
+	"github.com/sensu/sensu-go/backend/authorization/rbac"
+	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/transport"
 )
 
 // Authorization is an HTTP middleware that enforces authorization
@@ -41,6 +46,44 @@ func (a Authorization) Then(next http.Handler) http.Handler {
 			return
 		}
 
+		next.ServeHTTP(w, r.WithContext(ctx))
+	})
+}
+
+// BasicAuthorization performs basic authorization for event/entity creation via the agent websocket.
+func BasicAuthorization(next http.Handler, store store.Store) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		namespace := r.Header.Get(transport.HeaderKeyNamespace)
+		ctx := r.Context()
+		ctx = context.WithValue(ctx, corev2.NamespaceKey, namespace)
+
+		attrs := &authorization.Attributes{
+			APIGroup:     "core",
+			APIVersion:   "v2",
+			Namespace:    namespace,
+			Resource:     "events",
+			ResourceName: r.Header.Get(transport.HeaderKeyAgentName),
+			Verb:         "create",
+		}
+
+		err := getUser(ctx, attrs)
+		if err != nil {
+			writeErr(w, actions.NewErrorf(actions.PermissionDenied, "invalid user"))
+			return
+		}
+
+		auth := &rbac.Authorizer{
+			Store: store,
+		}
+		authorized, err := auth.Authorize(ctx, attrs)
+		if err != nil {
+			writeErr(w, actions.NewErrorf(actions.PermissionDenied, "error authorizing session"))
+			return
+		}
+		if !authorized {
+			writeErr(w, actions.NewErrorf(actions.PermissionDenied, "session is unauthorized"))
+			return
+		}
 		next.ServeHTTP(w, r.WithContext(ctx))
 	})
 }